fix: use duration value set from cli enable retrieval from secretStore for both cred types

+++ internal/config/config.go
@@ -19,5 +19,5 @@ type SamlConfig struct {
 	ProviderUrl  string
 	PrincipalArn string
 	AcsUrl       string
-	Duration     int
+	Duration     int64
 }
